#19 - remove NodeJS functionality
